+++ src/Bozboz/MediaLibrary/Decorators/MediaAdminDecorator.php
@@ -1,17 +1,25 @@
 <?php namespace Bozboz\MediaLibrary\Decorators;
 
 use Bozboz\Admin\Decorators\ModelAdminDecorator;
+use Bozboz\Admin\Fields\BelongsToManyField;
+use Bozboz\Admin\Fields\ListField;
+use Bozboz\Admin\Fields\SelectField;
 use Bozboz\Admin\Fields\TextField;
 use Bozboz\Admin\Reports\Filters\ArrayListingFilter;
+use Bozboz\Admin\Reports\Filters\MultiOptionListingFilter;
 use Bozboz\Admin\Reports\Filters\SearchListingFilter;
-
+use Bozboz\MediaLibrary\Decorators\TagAdminDecorator;
 use Bozboz\MediaLibrary\Fields\MediaField;
 use Bozboz\MediaLibrary\Models\Media;
 
 class MediaAdminDecorator extends ModelAdminDecorator
 {
-	public function __construct(Media $media)
+	protected $tags;
+
+	public function __construct(Media $media, TagAdminDecorator $tags)
 	{
+		$this->tags = $tags;
+
 		parent::__construct($media);
 	}
 
@@ -38,7 +46,11 @@ class MediaAdminDecorator extends ModelAdminDecorator
 			new TextField('caption'),
 			new MediaField($instance, array(
 				'name' => 'filename'
-			))
+			)),
+			new BelongsToManyField($this->tags, $instance->tags(), [
+				'key' => 'name',
+				'data-tags' => 'true'
+			])
 		);
 	}
 
@@ -47,18 +59,18 @@ class MediaAdminDecorator extends ModelAdminDecorator
 		return 'Media';
 	}
 
+	public function getListRelations()
+	{
+		return [
+			'tags' => 'name'
+		];
+	}
+
 	public function getListingFilters()
 	{
 		return [
 			new ArrayListingFilter('access', $this->getAccessOptions(), function($builder, $value) {
-				switch ($value) {
-					case Media::ACCESS_PUBLIC:
-						$builder->where('private', 0);
-					break;
-					case Media::ACCESS_PRIVATE:
-						$builder->where('private', 1);
-					break;
-				}
+				$builder->scope($value);
 			}, Media::ACCESS_BOTH),
 			new ArrayListingFilter('type', $this->getTypeOptions(), function($builder, $value) {
 				if ($value) {
@@ -66,6 +78,7 @@ class MediaAdminDecorator extends ModelAdminDecorator
 				}
 			}),
 			new SearchListingFilter('search', ['filename', 'caption']),
+			new MultiOptionListingFilter('tags', $this->model->tags()->getModel()->lists('name', 'id')),
 		];
 	}
 
@@ -80,6 +93,6 @@ class MediaAdminDecorator extends ModelAdminDecorator
 
 	private function getTypeOptions()
 	{
-		return ['All'] + array_map('ucwords', Media::groupBy('type')->lists('type', 'type'));
+		return ['All'] + array_map('ucwords', $this->model->groupBy('type')->lists('type', 'type'));
 	}
 }