+++ changelog.md
@@ -6,6 +6,22 @@
 -   Add $limit argument to ModelAdminDecorator::getListingModels
 
 
+## Version 1.2.0 (FUTURE)
+
+### Added
+-   Add tags to Media Library
+-   Add MultiOptionListingFilter
+-   Added messages functionality to Validator service
+-   View can now be overridden in constructor of Report class
+
+### Changed
+-   $editRules and passesEdit() method on Validator changed to $updateRules and passesUpdate()
+
+### Deprecated
+-   Deprecate Services\Validator::passesEdit() method
+-   Deprecate Reports\Report::overrideView() method
+
+
 ## Version 1.1.0 (2015-08-26)
 
 ### Added